requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.35.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-ads==24.1.0
openpyxl==3.1.2
